requirements.txt
logging
<<<<<<< HEAD
requests
=======
requests
pyodbc
# pydoc
azure-storage-blob
>>>>>>> c73fcbe (Clean start)
